third_party/blink/web_tests/wpt_internal/overscroll-gestures/parsing/overscroll-anchor-name-invalid.html

<!DOCTYPE html>
<meta charset="utf-8">

<title>Overscroll Anchor Name Parsing Test - invalid</title>
<meta name="assert" content="overscroll-anchor-name supports the full grammar 'none | <dashed-ident>#'.">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>

<script>
test_invalid_value("overscroll-anchor-name", "none --foo");
test_invalid_value("overscroll-anchor-name", "-foo");
test_invalid_value("overscroll-anchor-name", "foo");
test_invalid_value("overscroll-anchor-name", "foo --bar");
test_invalid_value("overscroll-anchor-name", "--foo bar");
test_invalid_value("overscroll-anchor-name", "--foo none");
</script>
